struct serial_rs485: Add padding
Typically the calls into Linux kernel are binary-compatible
across different kernel versions.
In order to be future proof, declare the padding words in
struct serial_rs485 as reserved.
They should be set to 0 (this is already done because we are
using callocMustSucceed()).
While at it, make more clear that we have 8 words of 32 bits
unsigned by using epicsUInt32.