Add new traceInfoMask bit ASYN_TRACEINFO_NOPATH
When the file/line printing is enabled by the bit ASYN_TRACEINFO_SOURCE,
file names are printed with help of the __FILE__ macro.
Due to our build system, the compiler is invoked with file names
like "../devMotorAsyn.c", and this is what we find in the IOC log file later.

However, the "../" is no useful information, so make it possible to
skip the "path-component" from file before the printing is done.

This will shorten the IOC log a little bit, and make it easier to read.